+++ android/app/src/main/kotlin/com/dunnowsoftware/GarageAAtoESP32/AndroidAutoState.kt
@@ -1,0 +1,5 @@
+package com.dunnowsoftware.GarageAAtoESP32
+
+object AndroidAutoState {
+    @Volatile var isConnected: Boolean = false
+}