+++ android/app/src/main/kotlin/com/dunnowsoftware/GarageAAtoESP32/geofence/BootReceiver.kt
@@ -3,7 +3,6 @@ package com.dunnowsoftware.GarageAAtoESP32.geofence
 import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
-
 class BootReceiver : BroadcastReceiver() {
     override fun onReceive(context: Context, intent: Intent) {
         if (intent.action == Intent.ACTION_BOOT_COMPLETED) {